Access taxon_concept data from CITES species+ API

Queries CITES Species+ API using an authentication token. The query string filters CITES/Species+ data by taxon concept (e.g. species, genus, class).

sppplus_taxonconcept(query_taxon, appendix_only = FALSE, token = NULL)

Arguments

query_taxon
a character string containing the query (e.g. species). Scientific taxa only.

appendix_only
a logical. If TRUE then taxon identifier and the CITES appendix information are the only data returned (default is FALSE).

token
a character string containing the authentification token, see https://api.speciesplus.net/documentation. Default is set to NULL and requires the environment variable SPPPLUS_TOKEN to be set directly in Renviron. Alternatively, sppplus_login() can be used to set SPPPLUS_TOKEN for the current session.

Value

If appendix_only is TRUE, then a data frame with a species' taxon id and CITES appendix information is returned. Otherwise, a list of objects of class data.table with all CITES Species+ taxon_concept information is returned. Importantly enough, this functions returns the taxon concept identifier that is required by the taxon_*() functions.

References

https://api.speciesplus.net/documentation/v1/taxon_concepts/index.html

Examples

# Not run:
# res1 <- sppplus_taxonconcept(query_taxon = 'Loxodonta africana')
# res2 <- sppplus_taxonconcept(query_taxon = 'Loxodonta africana', appendix_only = TRUE)
# res3 <- sppplus_taxonconcept(query_taxon = 'Amazilia versicolor')